samples/net/udp/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
bzip2=1.0.8=hcfcfb64_5
ca-certificates=2024.2.2=h56e8100_0
libexpat=2.6.2=h63175ca_0
libffi=3.4.2=h8ffe710_5
libsqlite=3.45.2=hcfcfb64_0
libzlib=1.2.13=hcfcfb64_5
openssl=3.2.1=hcfcfb64_0
pip=24.0=pyhd8ed1ab_0
pyserial=3.5=pyhd8ed1ab_0
python=3.12.2=h2628c8c_0_cpython
python_abi=3.12=4_cp312
pywin32=306=py312h53d5487_2
setuptools=69.2.0=pyhd8ed1ab_0
tk=8.6.13=h5226925_1
tzdata=2024a=h0c530f3_0
ucrt=10.0.22621.0=h57928b3_0
vc=14.3=hcf57466_18
vc14_runtime=14.38.33130=h82b7239_18
vs2015_runtime=14.38.33130=hcb4865c_18
wheel=0.42.0=pyhd8ed1ab_0
xz=5.2.6=h8d14728_0
